Fix type in Range attribute: change from double to int
An exception was thrown while loading the test.
System.InvalidOperationException: A value of type System.Int32 (1) cannot be passed to a parameter of type System.Double.

+++ TddCourse.Tests.Unit/Part08_ParameterizedTests/Parameterized_Range.cs
@@ -7,7 +7,7 @@ namespace TddCourse.Tests.Unit.Part08_ParameterizedTests
     {
         [Test]
         public void Divide_DividendIsZero_ReturnsQuotientEqualToZero(
-            [Range(@from: 1, to: 5, step: 1)] double divisor)
+            [Range(@from: 1, to: 5, step: 1)] int divisor)
         {
             var calc = new Calculator();
             float quotient = calc.Divide(0, divisor);
